PAY-65: Add status at interface and trait + Adjust models to use the new trait
